refactor: state를 분리하고 Main 컴포넌트에서 Restaurants 컴포넌트로 이름을 변경하다
